requirements.txt
praw
python-dotenv
google-genai
elevenlabs
yt_dlp
moviepy
pydub
pysrt
PIL
cleantext